backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20
supabase==2.3.4
pypdf==4.0.1
python-multipart==0.0.6
pydantic==2.5.3
